diffuse_vae_310/pyproject.toml
[project]
name = "dgm-2024-vae-diffusion"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Seungbin Yim", email = "ysb06@hotmail.com" }]
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "lightning>=2.4.0",
    "matplotlib>=3.8.0",
    "wandb>=0.18.5",
    "click>=8.1.7",
    "hydra-core>=1.3.2",
    "scipy>=1.13.1",
    "joblib>=1.4.2",
    "scikit-learn>=1.5.2",
    "lmdb>=1.5.1",
    "torch-fidelity>=0.3.0"
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm]
distribution = true

[dependency-groups]
test = ["pytest>=8.3.3"]
